requirements.txt
ansicon==1.89.0
arrow==1.2.3
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
blessed==1.20.0
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
DateTime==5.1
dj-database-url==2.0.0
Django==4.2.1
django-filter==23.2
django-heroku==0.3.1
django-picklefield==3.1
django-q==1.3.9
exceptiongroup==1.1.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
jinxed==1.2.0
outcome==1.2.0
psycopg2==2.9.6
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==3.5.3
requests==2.30.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
whitenoise==6.4.0
wsproto==1.2.0
zope.interface==6.0
